refactor: reorganize partition code into partition/ module with three submodules
Split partition code into a clear module hierarchy:

  kernel/src/partition/
      mod.rs              -- re-exports, data flow diagram
      hive.rs             -- Hive-style path percent-encoding (public)
      serialization.rs    -- Scalar -> String for AddFile.partitionValues (public)
      validation.rs       -- key completeness + type checking (pub(crate))

serialize_partition_value is now public for custom engines that bypass
DefaultEngine and need to build their own AddFile.partitionValues maps
(matching Java kernel's PartitionUtils.serializePartitionValue).
