+++ CLAUDE/architecture.md
@@ -130,7 +130,8 @@ all returned batches -- the engine may split a single file across multiple batch
   `column_expr!` macro
 - `kernel/src/transforms/` -- generic recursive transforms (`ExpressionTransform`,
   `SchemaTransform`)
-- `kernel/src/hive_partition.rs` -- Hive-style partition path encoding utilities
+- `kernel/src/partition/` -- partition utilities: `hive` (Hive-style path encoding),
+  `serialization` (Scalar-to-string for `partitionValues`), `validation` (key/type checks)
 - `kernel/src/checkpoint/` -- checkpoint writing (V1 and V2 single-file classic-named)
 - `kernel/src/table_configuration.rs` -- table metadata, properties, feature management
 - `kernel/src/table_features/` -- protocol feature definitions, `TableFeature` enum
