+++ kernel/src/lib.rs
@@ -96,11 +96,11 @@ pub(crate) mod crc;
 pub mod engine_data;
 pub mod error;
 pub mod expressions;
-pub mod hive_partition;
 mod log_compaction;
 mod log_path;
 mod log_reader;
 pub mod metrics;
+pub mod partition;
 pub mod scan;
 pub mod schema;
 pub mod snapshot;
